Disabled outlier detection by default.
Outlier detection does not yet work reliably and is therefore disabled by default. This is due to the residual function which currently is hardcoded and only subtracts the y values (img y value - fit y value). Even if the calculated point is very close to the curve, the y distance can still be quite big in contrast to the shortest distance to the curve. This is especially true for steep curves like Gaussian curves. The residual function will have to be improved by measuring the actual distance to the curve.